fix: resolve remaining Qt6 build issues
- Fix narrowing conversion from qreal to int in CodeEditorSideBar.cpp
- Fix Windows KSyntaxHighlighting build by building SyntaxHighlightingData
  target separately before the main install target to ensure proper
  dependency ordering for resource compilation

+++ src/Editor/CodeEditorSideBar.cpp
@@ -63,7 +63,7 @@ void CodeEditorSidebar::mouseReleaseEvent(QMouseEvent *event)
 {
     if (event->position().x() >= width() - m_codeEditor->fontMetrics().lineSpacing())
     {
-        auto block = m_codeEditor->blockAtPosition(event->position().y());
+        auto block = m_codeEditor->blockAtPosition(static_cast<int>(event->position().y()));
         if (!block.isValid() || !Editor::CodeEditor::isFoldable(block))
             return;
         m_codeEditor->toggleFold(block);
